project/education/requirements.txt
amqp==5.0.6
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.3.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
catboost==1.2
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
comm==0.1.3
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.11.0
DAWG-Python==0.7.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docopt==0.6.2
drf-yasg==1.20.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
fonttools==4.40.0
fqdn==1.5.1
graphviz==0.20.1
idna==2.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
install==1.3.4
intervaltree==3.1.0
ipykernel==6.23.2
ipymarkup==0.9.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.2.0
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.6.3
jupyter-server==2.6.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kombu==5.0.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
natasha==1.5.0
navec==0.10.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.24.3
oauthlib==3.1.0
openapi-codec==1.3.2
overrides==7.3.1
packaging==20.9
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.6.0
plotly==5.15.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.15.1
PyJWT==2.1.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.4
python-json-logger==2.0.7
python3-openid==3.2.0
pytils==0.3
pytz==2021.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
razdel==0.5.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
simplejson==3.17.2
six==1.15.0
sklearn==0.0.post5
slovnet==0.6.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.1
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.6.3
tzdata==2023.3
uri-template==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
widgetsnbextension==4.0.7
yargy==0.15.1
zipp==3.15.0